+++ examples/4-fill-table.lua
@@ -1,67 +1,67 @@
 local ft = require "fort"
 
 local function use_printf()
-    local table = ft.create_table()
+    local ftable = ft.create_table()
 
-    ft.set_cell_prop(table, 0, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
-    ft.printf_ln(table, "N|Planet|Speed, km/s")
+    ft.set_cell_prop(ftable, 1, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
+    ft.printf_ln(ftable, "N|Planet|Speed, km/s")
 
-    ft.printf_ln(table, "%d|%s|%5.2f", 1, "Mercury", 47.362)
-    ft.printf_ln(table, "%d|%s|%5.2f", 2, "Venus", 35.02)
-    ft.printf_ln(table, "%d|%s|%5.2f", 3, "Earth", 29.78)
+    ft.printf_ln(ftable, "%d|%s|%5.2f", 1, "Mercury", 47.362)
+    ft.printf_ln(ftable, "%d|%s|%5.2f", 2, "Venus", 35.02)
+    ft.printf_ln(ftable, "%d|%s|%5.2f", 3, "Earth", 29.78)
 
-    print(ft.to_string(table))
+    print(ftable)
 end
 
 local function use_print()
-    local table = ft.create_table()
+    local ftable = ft.create_table()
 
-    ft.set_cell_prop(table, 0, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
-    ft.print_ln(table, "N|Planet|Speed, km/s")
+    ft.set_cell_prop(ftable, 1, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
+    ft.print_ln(ftable, "N|Planet|Speed, km/s")
 
-    ft.print_ln(table, "1|Mercury|47.362")
-    ft.print_ln(table, "2|Venus|35.02")
-    ft.print_ln(table, "3|Earth|29.78")
+    ft.print_ln(ftable, "1|Mercury|47.362")
+    ft.print_ln(ftable, "2|Venus|35.02")
+    ft.print_ln(ftable, "3|Earth|29.78")
 
-    print(ft.to_string(table))
+    print(ftable)
 end
 
 local function use_write()
-    local table = ft.create_table()
+    local ftable = ft.create_table()
 
-    ft.set_cell_prop(table, 0, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
-    ft.write_ln(table, "N", "Planet", "Speed, km/s")
+    ft.set_cell_prop(ftable, 1, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
+    ft.write_ln(ftable, "N", "Planet", "Speed, km/s")
 
-    ft.write_ln(table, "1", "Mercury", "47.362")
-    ft.write_ln(table, "2", "Venus", "35.02")
-    ft.write_ln(table, "3", "Earth", "29.78")
+    ft.write_ln(ftable, "1", "Mercury", "47.362")
+    ft.write_ln(ftable, "2", "Venus", "35.02")
+    ft.write_ln(ftable, "3", "Earth", "29.78")
 
-    print(ft.to_string(table))
+    print(ftable)
 end
 
 local function use_row_write()
-    local table = ft.create_table()
+    local ftable = ft.create_table()
 
-    ft.set_cell_prop(table, 0, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
-    ft.row_write_ln(table, {"N", "Planet", "Speed, km/s"})
+    ft.set_cell_prop(ftable, 1, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
+    ft.row_write_ln(ftable, {"N", "Planet", "Speed, km/s"})
 
-    ft.row_write_ln(table, {"1", "Mercury", "47.362"})
-    ft.row_write_ln(table, {"2", "Venus", "35.02"})
-    ft.row_write_ln(table, {"3", "Earth", "29.78"})
+    ft.row_write_ln(ftable, {"1", "Mercury", "47.362"})
+    ft.row_write_ln(ftable, {"2", "Venus", "35.02"})
+    ft.row_write_ln(ftable, {"3", "Earth", "29.78"})
 
-    print(ft.to_string(table))
+    print(ftable)
 end
 
 local function use_table_write()
-    local table = ft.create_table()
+    local ftable = ft.create_table()
 
-    ft.set_cell_prop(table, 0, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
-    ft.table_write_ln(table, {
+    ft.set_cell_prop(ftable, 1, ft.ANY_COLUMN, ft.CPROP_ROW_TYPE, ft.ROW_HEADER)
+    ft.table_write_ln(ftable, {
         {"N", "Planet", "Speed, km/s"}, {"1", "Mercury", "47.362"},
         {"2", "Venus", "35.02"}, {"3", "Earth", "29.78"}
     })
 
-    print(ft.to_string(table))
+    print(ftable)
 
 end