+++ examples/9-table-grid.lua
@@ -9,4 +9,14 @@ ft.write_ln(ftable, "1", "Ricciardo", "1:25.945", "222.128")
 ft.write_ln(ftable, "2", "Hamilton", "1:26.373", "221.027")
 ft.write_ln(ftable, "3", "Verstappen", "1:26.469", "220.782")
 
-print(ftable)
+--- create a table grid
+local fgrid = ft.create_table()
+ft.set_border_style(fgrid, ft.EMPTY_STYLE)
+
+local ftable2 = ft.copy(ftable)
+
+ft.write_ln(ftable2, "4", "?", "?", "?")
+
+ft.table_write(fgrid, {{ftable, ftable2, ftable}, {ftable, "", ftable}})
+
+print(fgrid)