+++ src/fort.lua
@@ -1,9 +1,14 @@
 --- Fort Module
--- @classmod fort
+-- @module fort
 -- @pragma nostrip
-local fort = require "cfort"
+local cfort = require "cfort"
 
-local fortc = {}
+---@class fort
+local fort = {}
+
+-- enable new functions to be accessed in ftable object
+setmetatable(cfort, {__index = fort})
+setmetatable(fort, fort)
 
 local function split(inputstr, sep)
     if sep == nil then sep = "%s" end
@@ -14,14 +19,18 @@ local function split(inputstr, sep)
     return t
 end
 
+---Create a new formatted table
+---@return fort new formatted table
+function fort.create_table() return cfort.create_table() end
+
 ---Alias of @{fort.create_table}
----@function fort.new
 fort.new = fort.create_table
 
 ---Alias of @{fort.create_table}
----@within Metamethods
+fort.create = fort.create_table
+
+---Alias of @{fort.create_table}
 fort.__call = fort.new
-setmetatable(fort, fort)
 
 ---Control the default separator for @{printf} and @{printf_ln} functions.
 ---(Defaults to '|')
@@ -31,9 +40,8 @@ fort.default_separator = "|"
 ---
 ---Splits cells on fort.default_separator. Uses lua @{string.format} to
 ---format the text.
----@string row_format row to write with formatting
----@param ... any format arguments
----@within Methods
+---@param row_format string row to write with formatting
+---@vararg ... any format arguments
 function fort:printf(row_format, ...)
     local formatted_text = string.format(row_format, ...)
     local row = split(formatted_text, fort.default_separator)
@@ -44,9 +52,8 @@ end
 ---
 ---Splits cells on fort.default_separator. Uses lua @{string.format} to
 ---format the text.
----@string row_format row to write with formatting
----@param ... any format arguments
----@within Methods
+---@param row_format string row to write with formatting
+---@vararg ... any format arguments
 function fort:printf_ln(row_format, ...)
     local formatted_text = string.format(row_format, ...)
     local row = split(formatted_text, fort.default_separator)
@@ -56,9 +63,8 @@ end
 ---Write a row.
 ---Use a single string to write a row. Allows using a custom cell separator
 ---without modifying the global separator.
----@string row_text
----@string[opt=fort.default_separator] sep cell separator
----@within Methods
+---@param row_text string
+---@param sep string? sep cell separator
 function fort:print(row_text, sep)
     sep = sep or fort.default_separator
     local row = split(row_text, sep)
@@ -68,9 +74,8 @@ end
 ---Write a row and go to the next line.
 ---Use a single string to write a row. Allows using a custom cell separator
 ---without modifying the global separator.
----@string row_text
----@string[opt=fort.default_separator] sep cell separator
----@within Methods
+---@param row_text string
+---@param sep string? sep cell separator
 function fort:print_ln(row_text, sep)
     sep = sep or fort.default_separator
     local row = split(row_text, sep)
@@ -78,9 +83,8 @@ function fort:print_ln(row_text, sep)
 end
 
 ---Write a 2d array of strings to the ftable.
----@tparam {{string}} data_table 2d array of strings to write, can be jagged.
----@bool[opt=false] colalign align to the @{cur_col} at the start
----@within Methods
+---@param data_table string[][] data_table 2d array of strings to write, can be jagged.
+---@param colalign boolean?  align to the @{cur_col} at the start
 function fort:table_write(data_table, colalign)
     colalign = colalign or false
     local rows = #data_table
@@ -98,9 +102,8 @@ function fort:table_write(data_table, colalign)
 end
 
 ---Write a 2d array of strings to the ftable and go to next line.
----@tparam {{string}} data_table 2d array of strings to write, can be jagged.
----@bool[opt=false] colalign align to the @{cur_col} at the start
----@within Methods
+---@param data_table string[][] 2d array of strings to write, can be jagged.
+---@param colalign boolean? align to the @{cur_col} at the start
 function fort:table_write_ln(data_table, colalign)
     local cur_col = fort.cur_col(self)
     colalign = colalign or false
@@ -113,136 +116,130 @@ function fort:table_write_ln(data_table, colalign)
 end
 
 ---Add a dividing separtor line at the current row.
----@function fort:add_separator
----@within Methods
+function fort:add_separator() cfort.add_separator(self) end
 
 ---Completely Copy a table
----@function fort:copy_table
----@within Methods
----@treturn ftable copied table
+---@return fort copied table
+function fort:copy_table() return cfort.copy_table(self) end
 
 ---Alias of @{fort.copy_table}
----@function fort:copy
----@within Methods
----@treturn ftable copied table
 fort.copy = fort.copy_table
 
----Create a new formatted table
----@function fort.create_table
----@treturn ftable new formatted table
-
 ---Get the current column
----@function fort:cur_col
----@within Methods
----@treturn number
+---@return integer
+function fort:cur_col() return cfort.cur_col(self) end
 
 ---Get the current row
----@function fort:cur_row
----@within Methods
----@treturn number
+---@return integer
+function fort:cur_row() return cfort.cur_row(self) end
 
 ---Erase a rectangular range of data from the ftable
----@function fort:erase_range
----@within Methods
----@number top_left_row
----@number top_left_col
----@number bottom_right_row
----@number bottom_right_col
-
----Check if ftable is empty
----@function fort:is_empty
----@within Methods
----@treturn boolean
-
----Go to the next line (row)
----@function fort:ln
----@within Methods
-
----Get the number of rows in the ftable
----@function fort:row_count
----@within Methods
----@treturn number
-
-fortc.row_write = fort.row_write
-
----Write a row of data.
----@tparam {string} row row of strings to write
----@within Methods
-function fort:row_write(row)
-    local stringified_row = {}
-    for i, v in ipairs(row) do stringified_row[i] = tostring(v) end
-    fortc.row_write(self, stringified_row)
+---@param top_left_row integer
+---@param top_left_col integer
+---@param bottom_right_row integer
+---@param bottom_right_col integer
+function fort:erase_range(top_left_row, top_left_col, bottom_right_row,
+                          bottom_right_col)
+    return cfort.erase_range(self, top_left_row, top_left_col, bottom_right_row,
+                             bottom_right_col)
 end
 
-fortc.row_write_ln = fort.row_write_ln
----Write a row of data and go to the next line.
----@tparam {string} row row of strings to write
----@within Methods
-function fort:row_write_ln(row)
+---Check if ftable is empty
+---@return boolean
+function fort:is_empty() return cfort.is_empty(self) end
+
+---Go to the next line (row)
+function fort:ln() return cfort.ln(self) end
+
+---Get the number of rows in the ftable
+---@return integer
+function fort:row_count() return cfort.row_count(self) end
+
+---Get the number of columns in the ftable
+---@return integer
+function fort:col_count() return cfort.col_count(self) end
+
+---Write a row of data.
+---@param row any[] row of data to write (uses tostring)
+function fort:row_write(row)
+    assert(type(row) == "table", "Expected table for arg 1")
     local stringified_row = {}
     for i, v in ipairs(row) do stringified_row[i] = tostring(v) end
-    fortc.row_write_ln(self, stringified_row)
+    cfort.row_write(self, stringified_row)
+end
+
+---Write a row of data and go to the next line.
+---@param row any[] row of data to write (uses tostring)
+function fort:row_write_ln(row)
+    assert(type(row) == "table", "Expected table for arg 1")
+    local stringified_row = {}
+    for i, v in ipairs(row) do stringified_row[i] = tostring(v) end
+    cfort.row_write_ln(self, stringified_row)
 end
 
 ---Write a row
----@param ... string strings to write in the row
----@within Methods
-function fort:write(...) fortc.row_write(self, {...}) end
+---@vararg any data to write in the row (uses tostring)
+function fort:write(...) fort.row_write(self, {...}) end
 ---Write a row and go to the next line
----@param ... string strings to write in the row
----@within Methods
+---@vararg any data to write in the row
 function fort:write_ln(...) fort.row_write_ln(self, {...}) end
 
 ---Set the border style of the ftable.
----@function fort:set_border_style
----@within Methods
----@userdata style from available styles @{BASIC_STYLE}
+---@param style fort.BorderStyle available styles @{BASIC_STYLE}
+function fort:set_border_style(style) cfort.set_border_style(self, style) end
 
 ---Set the cell property of the ftable.
----@function fort:set_cell_prop
----@within Methods
----@number row the row to set, can also use @{ANY_ROW}/@{CUR_ROW}
----@number col the column to set, can also use @{ANY_COLUMN}/@{CUR_COLUMN}
----@number property the property to set
----@number value value to set
+---@param row integer the row to set, can also use @{ANY_ROW}/@{CUR_ROW}
+---@param col integer column to set, can also use @{ANY_COLUMN}/@{CUR_COLUMN}
+---@param property fort.CellProperty property to set
+---@param value integer to set
+function fort:set_cell_prop(row, col, property, value)
+    cfort.set_cell_prop(self, row, col, property, value)
+end
 
 ---Set a cell's horizontal span in the ftable.
----@function fort:set_cell_span
----@within Methods
----@number row the row to set. DO NOT USE @{ANY_ROW}/@{CUR_ROW}
----@number col the column to set. DO NOT USE @{ANY_COLUMN}/@{CUR_COLUMN}
----@number span how many columns the cell should span
+---@param row integer the row to set. DO NOT USE @{ANY_ROW}/@{CUR_ROW}
+---@param col integer the column to set. DO NOT USE @{ANY_COLUMN}/@{CUR_COLUMN}
+---@param span integer how many columns the cell should span
+function fort:set_cell_span(row, col, span)
+    cfort.set_cell_span(self, row, col, span)
+end
 
 ---Set the current cell position.
----@function fort:set_cur_cell
----@within Methods
----@number row the row to set. DO NOT USE @{ANY_ROW}/@{CUR_ROW}
----@number col the column to set. DO NOT USE @{ANY_COLUMN}/@{CUR_COLUMN}
+---@param row integer the row to set. DO NOT USE @{ANY_ROW}/@{CUR_ROW}
+---@param col integer the column to set. DO NOT USE @{ANY_COLUMN}/@{CUR_COLUMN}
+function fort:set_cur_cell(row, col) cfort.set_cur_cell(self, row, col) end
 
 ---Set the default border style for new tables.
----@function fort.set_default_border_style
----@userdata style from available styles @{BASIC_STYLE}
+---@param style fort.BorderStyle available styles @{BASIC_STYLE}
+function fort.set_default_border_style(style)
+    cfort.set_default_border_style(style)
+end
 
 ---Set a default cell property for new tables.
----@function fort.set_default_cell_prop
----@number property the property to set
----@number value value to set
+---@param property fort.CellProperty the property to set
+---@param value integer value to set
+function fort.set_default_cell_prop(property, value)
+    cfort.set_default_cell_prop(property, value)
+end
 
 ---Set a default table property for new tables.
----@number property the property to set
----@number value value to set
----@function fort.set_default_tbl_prop
+---@param property fort.TableProperty the property to set
+---@param value integer value to set
+function fort.set_default_tbl_prop(property, value)
+    cfort.set_default_tbl_prop(property, value)
+end
 
 ---Set a table property.
----@function fort:set_tbl_prop
----@within Methods
----@number property the property to set
----@number value value to set
+---@param property fort.TableProperty the property to set
+---@param value integer value to set
+function fort:set_tbl_prop(property, value)
+    cfort.set_tbl_prop(self, property, value)
+end
 
 ---Generate the string version of the ftable.
----@function fort:to_string
----@within Methods
----@treturn string formatted table string
+---@return string formatted table string
+function fort:to_string() return cfort.to_string(self) end
 
 ---Cell Text Align.
 -- Indicate the text alignment inside a cell.
@@ -250,12 +247,12 @@ function fort:write_ln(...) fort.row_write_ln(self, {...}) end
 -- Can be used with @{set_cell_prop}/@{set_default_cell_prop}
 -- @section cell-text-align
 
---- align text center
-fort.ALIGNED_CENTER = fort.ALIGNED_CENTER
---- align text left
-fort.ALIGNED_LEFT = fort.ALIGNED_LEFT
---- align text right
-fort.ALIGNED_RIGHT = fort.ALIGNED_RIGHT
+---@type integer
+fort.ALIGNED_CENTER = cfort.ALIGNED_CENTER
+---@type integer
+fort.ALIGNED_LEFT = cfort.ALIGNED_LEFT
+---@type integer
+fort.ALIGNED_RIGHT = cfort.ALIGNED_RIGHT
 
 ---Cell Selectors.
 -- Special flags that select a cell
@@ -263,14 +260,18 @@ fort.ALIGNED_RIGHT = fort.ALIGNED_RIGHT
 -- Can be used with @{set_cell_prop}
 -- @section cell-selector
 
+---@type integer
 --- Select all columns
-fort.ANY_COLUMN = fort.ANY_COLUMN
+fort.ANY_COLUMN = cfort.ANY_COLUMN
+---@type integer
 --- Select all rows
-fort.ANY_ROW = fort.ANY_ROW
+fort.ANY_ROW = cfort.ANY_ROW
+---@type integer
 --- Select the current column
-fort.CUR_COLUMN = fort.CUR_COLUMN
+fort.CUR_COLUMN = cfort.CUR_COLUMN
+---@type integer
 --- Select the current row
-fort.CUR_ROW = fort.CUR_ROW
+fort.CUR_ROW = cfort.CUR_ROW
 
 ---Table Border Style.
 -- Border styling of the table.
@@ -280,154 +281,156 @@ fort.CUR_ROW = fort.CUR_ROW
 -- to see how each border style looks.
 -- @section table-border-style
 
----Border style
-fort.BASIC_STYLE = fort.BASIC_STYLE
----Border style
-fort.BASIC2_STYLE = fort.BASIC2_STYLE
----Border style
-fort.BOLD_STYLE = fort.BOLD_STYLE
----Border style
-fort.BOLD2_STYLE = fort.BOLD2_STYLE
----Border style
-fort.DOT_STYLE = fort.DOT_STYLE
----Border style
-fort.DOUBLE_STYLE = fort.DOUBLE_STYLE
----Border style
-fort.DOUBLE2_STYLE = fort.DOUBLE2_STYLE
----Border style
-fort.EMPTY_STYLE = fort.EMPTY_STYLE
----Border style
-fort.EMPTY2_STYLE = fort.EMPTY2_STYLE
----Border style
-fort.FRAME_STYLE = fort.FRAME_STYLE
----Border style
-fort.NICE_STYLE = fort.NICE_STYLE
----Border style
-fort.PLAIN_STYLE = fort.PLAIN_STYLE
----Border style
-fort.SIMPLE_STYLE = fort.SIMPLE_STYLE
----Border style
-fort.SOLID_ROUND_STYLE = fort.SOLID_ROUND_STYLE
----Border style
-fort.SOLID_STYLE = fort.SOLID_STYLE
+---@class fort.BorderStyle
+
+---@type fort.BorderStyle
+fort.BASIC_STYLE = cfort.BASIC_STYLE
+---@type fort.BorderStyle
+fort.BASIC2_STYLE = cfort.BASIC2_STYLE
+---@type fort.BorderStyle
+fort.BOLD_STYLE = cfort.BOLD_STYLE
+---@type fort.BorderStyle
+fort.BOLD2_STYLE = cfort.BOLD2_STYLE
+---@type fort.BorderStyle
+fort.DOT_STYLE = cfort.DOT_STYLE
+---@type fort.BorderStyle
+fort.DOUBLE_STYLE = cfort.DOUBLE_STYLE
+---@type fort.BorderStyle
+fort.DOUBLE2_STYLE = cfort.DOUBLE2_STYLE
+---@type fort.BorderStyle
+fort.EMPTY_STYLE = cfort.EMPTY_STYLE
+---@type fort.BorderStyle
+fort.EMPTY2_STYLE = cfort.EMPTY2_STYLE
+---@type fort.BorderStyle
+fort.FRAME_STYLE = cfort.FRAME_STYLE
+---@type fort.BorderStyle
+fort.NICE_STYLE = cfort.NICE_STYLE
+---@type fort.BorderStyle
+fort.PLAIN_STYLE = cfort.PLAIN_STYLE
+---@type fort.BorderStyle
+fort.SIMPLE_STYLE = cfort.SIMPLE_STYLE
+---@type fort.BorderStyle
+fort.SOLID_ROUND_STYLE = cfort.SOLID_ROUND_STYLE
+---@type fort.BorderStyle
+fort.SOLID_STYLE = cfort.SOLID_STYLE
 
 ---Cell Color.
 -- Color used for cell/content properties.
 -- @section cell-color
 
----Color
-fort.COLOR_BLACK = fort.COLOR_BLACK
----Color
-fort.COLOR_BLUE = fort.COLOR_BLUE
----Color
-fort.COLOR_CYAN = fort.COLOR_CYAN
----Color
-fort.COLOR_DARK_GRAY = fort.COLOR_DARK_GRAY
----Color
-fort.COLOR_DEFAULT = fort.COLOR_DEFAULT
----Color
-fort.COLOR_GREEN = fort.COLOR_GREEN
----Color
-fort.COLOR_LIGHT_BLUE = fort.COLOR_LIGHT_BLUE
----Color
-fort.COLOR_LIGHT_CYAN = fort.COLOR_LIGHT_CYAN
----Color
-fort.COLOR_LIGHT_GRAY = fort.COLOR_LIGHT_GRAY
----Color
-fort.COLOR_LIGHT_GREEN = fort.COLOR_LIGHT_GREEN
----Color
-fort.COLOR_LIGHT_MAGENTA = fort.COLOR_LIGHT_MAGENTA
----Color
-fort.COLOR_LIGHT_RED = fort.COLOR_LIGHT_RED
----Color
-fort.COLOR_LIGHT_WHITE = fort.COLOR_LIGHT_WHYTE
----Color
-fort.COLOR_LIGHT_YELLOW = fort.COLOR_LIGHT_YELLOW
----Color
-fort.COLOR_MAGENTA = fort.COLOR_MAGENTA
----Color
-fort.COLOR_RED = fort.COLOR_RED
----Color
-fort.COLOR_YELLOW = fort.COLOR_YELLOW
+---@type integer
+fort.COLOR_BLACK = cfort.COLOR_BLACK
+---@type integer
+fort.COLOR_BLUE = cfort.COLOR_BLUE
+---@type integer
+fort.COLOR_CYAN = cfort.COLOR_CYAN
+---@type integer
+fort.COLOR_DARK_GRAY = cfort.COLOR_DARK_GRAY
+---@type integer
+fort.COLOR_DEFAULT = cfort.COLOR_DEFAULT
+---@type integer
+fort.COLOR_GREEN = cfort.COLOR_GREEN
+---@type integer
+fort.COLOR_LIGHT_BLUE = cfort.COLOR_LIGHT_BLUE
+---@type integer
+fort.COLOR_LIGHT_CYAN = cfort.COLOR_LIGHT_CYAN
+---@type integer
+fort.COLOR_LIGHT_GRAY = cfort.COLOR_LIGHT_GRAY
+---@type integer
+fort.COLOR_LIGHT_GREEN = cfort.COLOR_LIGHT_GREEN
+---@type integer
+fort.COLOR_LIGHT_MAGENTA = cfort.COLOR_LIGHT_MAGENTA
+---@type integer
+fort.COLOR_LIGHT_RED = cfort.COLOR_LIGHT_RED
+---@type integer
+fort.COLOR_LIGHT_WHITE = cfort.COLOR_LIGHT_WHYTE
+---@type integer
+fort.COLOR_LIGHT_YELLOW = cfort.COLOR_LIGHT_YELLOW
+---@type integer
+fort.COLOR_MAGENTA = cfort.COLOR_MAGENTA
+---@type integer
+fort.COLOR_RED = cfort.COLOR_RED
+---@type integer
+fort.COLOR_YELLOW = cfort.COLOR_YELLOW
 
----Cell Property.
--- @section cell-property
+---@class fort.CellProperty : number
 
----Cell Property
-fort.CPROP_CELL_BG_COLOR = fort.CPROP_CELL_BG_COLOR
----Cell Property
-fort.CPROP_CELL_TEXT_STYLE = fort.CPROP_CELL_TEXT_STYLE
----Cell Property
-fort.CPROP_CONT_BG_COLOR = fort.CPROP_CONT_BG_COLOR
----Cell Property
-fort.CPROP_CONT_FG_COLOR = fort.CPROP_CONT_FG_COLOR
----Cell Property
-fort.CPROP_CONT_TEXT_STYLE = fort.CPROP_CONT_TEXT_STYLE
----Cell Property
-fort.CPROP_EMPTY_STR_HEIGHT = fort.CPROP_EMPTY_STR_HEIGHT
----Cell Property
-fort.CPROP_MIN_WIDTH = fort.CPROP_MIN_WIDTH
----Cell Property
-fort.CPROP_ROW_TYPE = fort.CPROP_ROW_TYPE
----Cell Property
-fort.CPROP_TEXT_ALIGN = fort.CPROP_TEXT_ALIGN
----Cell Property
-fort.CPROP_TOP_PADDING = fort.CPROP_TOP_PADDING
----Cell Property
-fort.CPROP_LEFT_PADDING = fort.CPROP_LEFT_PADDING
----Cell Property
-fort.CPROP_BOTTOM_PADDING = fort.CPROP_BOTTOM_PADDING
----Cell Property
-fort.CPROP_RIGHT_PADDING = fort.CPROP_RIGHT_PADDING
+---@type fort.CellProperty
+fort.CPROP_CELL_BG_COLOR = cfort.CPROP_CELL_BG_COLOR
+---@type fort.CellProperty
+fort.CPROP_CELL_TEXT_STYLE = cfort.CPROP_CELL_TEXT_STYLE
+---@type fort.CellProperty
+fort.CPROP_CONT_BG_COLOR = cfort.CPROP_CONT_BG_COLOR
+---@type fort.CellProperty
+fort.CPROP_CONT_FG_COLOR = cfort.CPROP_CONT_FG_COLOR
+---@type fort.CellProperty
+fort.CPROP_CONT_TEXT_STYLE = cfort.CPROP_CONT_TEXT_STYLE
+---@type fort.CellProperty
+fort.CPROP_EMPTY_STR_HEIGHT = cfort.CPROP_EMPTY_STR_HEIGHT
+---@type fort.CellProperty
+fort.CPROP_MIN_WIDTH = cfort.CPROP_MIN_WIDTH
+---@type fort.CellProperty
+fort.CPROP_ROW_TYPE = cfort.CPROP_ROW_TYPE
+---@type fort.CellProperty
+fort.CPROP_TEXT_ALIGN = cfort.CPROP_TEXT_ALIGN
+---@type fort.CellProperty
+fort.CPROP_TOP_PADDING = cfort.CPROP_TOP_PADDING
+---@type fort.CellProperty
+fort.CPROP_LEFT_PADDING = cfort.CPROP_LEFT_PADDING
+---@type fort.CellProperty
+fort.CPROP_BOTTOM_PADDING = cfort.CPROP_BOTTOM_PADDING
+---@type fort.CellProperty
+fort.CPROP_RIGHT_PADDING = cfort.CPROP_RIGHT_PADDING
 
 ---Row Type.
 -- @section row-type
 
----Row Type
-fort.ROW_COMMON = fort.ROW_COMMON
----Row Type
-fort.ROW_HEADER = fort.ROW_HEADER
+---@type integer
+fort.ROW_COMMON = cfort.ROW_COMMON
+---@type integer
+fort.ROW_HEADER = cfort.ROW_HEADER
 
 ---Table Adding Strategy.
 -- @section table-adding-strategy
 
+---@type integer
 ---Insert new cells
-fort.STRATEGY_INSERT = fort.STRATEGY_INSERT
+fort.STRATEGY_INSERT = cfort.STRATEGY_INSERT
+---@type integer
 ---Replace current cells
-fort.STRATEGY_REPLACE = fort.STRATEGY_REPLACE
+fort.STRATEGY_REPLACE = cfort.STRATEGY_REPLACE
 
----Table Property.
--- @section table-property
+---@class fort.TableProperty : number
 
----Table Property
-fort.TPROP_ADDING_STRATEGY = fort.TPROP_ADDING_STRATEGY
----Table Property
-fort.TPROP_BOTTOM_MARGIN = fort.TPROP_BOTTOM_MARGIN
----Table Property
-fort.TPROP_LEFT_MARGIN = fort.TPROP_LEFT_MARGIN
----Table Property
-fort.TPROP_RIGHT_MARGIN = fort.TPROP_RIGHT_MARGIN
----Table Property
-fort.TPROP_TOP_MARGIN = fort.TPROP_TOP_MARGIN
+---@type fort.TableProperty
+fort.TPROP_ADDING_STRATEGY = cfort.TPROP_ADDING_STRATEGY
+---@type fort.TableProperty
+fort.TPROP_BOTTOM_MARGIN = cfort.TPROP_BOTTOM_MARGIN
+---@type fort.TableProperty
+fort.TPROP_LEFT_MARGIN = cfort.TPROP_LEFT_MARGIN
+---@type fort.TableProperty
+fort.TPROP_RIGHT_MARGIN = cfort.TPROP_RIGHT_MARGIN
+---@type fort.TableProperty
+fort.TPROP_TOP_MARGIN = cfort.TPROP_TOP_MARGIN
 
 ---Cell Text Style.
 -- @section cell-text-style
 
----Text Style
-fort.TSTYLE_BLINK = fort.TSTYLE_BLINK
----Text Style
-fort.TSTYLE_BOLD = fort.TSTYLE_BOLD
----Text Style
-fort.TSTYLE_DEFAULT = fort.TSTYLE_DEFAULT
----Text Style
-fort.TSTYLE_DIM = fort.TSTYLE_DIM
----Text Style
-fort.TSTYLE_HIDDEN = fort.TSTYLE_HIDDEN
----Text Style
-fort.TSTYLE_INVERTED = fort.TSTYLE_INVERTED
----Text Style
-fort.TSTYLE_ITALIC = fort.TSTYLE_ITALIC
----Text Style
-fort.TSTYLE_UNDERLINED = fort.TSTYLE_UNDERLINED
+---@type integer
+fort.TSTYLE_BLINK = cfort.TSTYLE_BLINK
+---@type integer
+fort.TSTYLE_BOLD = cfort.TSTYLE_BOLD
+---@type integer
+fort.TSTYLE_DEFAULT = cfort.TSTYLE_DEFAULT
+---@type integer
+fort.TSTYLE_DIM = cfort.TSTYLE_DIM
+---@type integer
+fort.TSTYLE_HIDDEN = cfort.TSTYLE_HIDDEN
+---@type integer
+fort.TSTYLE_INVERTED = cfort.TSTYLE_INVERTED
+---@type integer
+fort.TSTYLE_ITALIC = cfort.TSTYLE_ITALIC
+---@type integer
+fort.TSTYLE_UNDERLINED = cfort.TSTYLE_UNDERLINED
 
 return fort