+++ src/lfort.c
@@ -11,23 +11,22 @@
 #define lua_tbl_len(L, arg) (lua_objlen(L, arg))
 #endif
 
+// use integer in newer lua for response values
+#if LUA_VERSION_NUM < 503
+#define lua_pushinteger(L, n) lua_pushinteger(L, n)
+#endif
+
 #define FTABLEMETA "fort.ftable"
 #define FBORDERSTYLE "fort.border_style"
 
-#define ERR_CHECK(func)                                     \
-    do {                                                    \
-        int _error_code = func;                             \
-        if (_error_code != 0) {                             \
-            return luaL_error(L, ft_strerror(_error_code)); \
-        }                                                   \
+#define ERR_CHECK(func)                                                      \
+    do {                                                                     \
+        int _error_code = func;                                              \
+        if (_error_code != 0) {                                              \
+            return luaL_error(L, "fort: %s (%d) ", ft_strerror(_error_code), \
+                              _error_code);                                  \
+        }                                                                    \
     } while (0);
-
-static ft_table_t **get_fort_table(lua_State *L, int arg_num) {
-    if (!lua_isuserdata(L, arg_num)) {
-        return NULL;
-    }
-    return (ft_table_t **)lua_touserdata(L, arg_num);
-}
 
 static int ftable_destroy(lua_State *L) {
     ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
@@ -50,9 +49,8 @@ static void register_ftable(lua_State *L, ft_table_t *table) {
 
 static int lft_create_table(lua_State *L) {
     ft_table_t *table = ft_create_table();
-    if (table == NULL) {
-        return luaL_error(L, "Unable to create table");
-    }
+    ERR_CHECK(table != NULL ? FT_SUCCESS : FT_MEMORY_ERROR);
+
     register_ftable(L, table);
     return 1;
 }
@@ -60,6 +58,7 @@ static int lft_create_table(lua_State *L) {
 static int lft_copy_table(lua_State *L) {
     ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
     ft_table_t *new_table = ft_copy_table(*table);
+    ERR_CHECK((new_table != NULL) ? FT_SUCCESS : FT_GEN_ERROR);
 
     register_ftable(L, new_table);
     return 1;
@@ -75,21 +74,24 @@ static int lft_cur_row(lua_State *L) {
     ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
 
     int cur_row = ft_cur_row(*table);
-    lua_pushnumber(L, cur_row);
+    lua_pushinteger(L, cur_row + 1);
     return 1;
 }
 
 static int lft_cur_col(lua_State *L) {
     ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
     int cur_col = ft_cur_col(*table);
-    lua_pushnumber(L, cur_col);
+    lua_pushinteger(L, cur_col + 1);
     return 1;
 }
 
 static int lft_set_cur_cell(lua_State *L) {
     ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
-    size_t row = luaL_checknumber(L, 2);
-    size_t col = luaL_checknumber(L, 3);
+    size_t row = luaL_checknumber(L, 2) - 1;
+    size_t col = luaL_checknumber(L, 3) - 1;
+
+    luaL_argcheck(L, lua_tonumber(L, 2) > 0, 2, "must be > 0");
+    luaL_argcheck(L, lua_tonumber(L, 3) > 0, 3, "must be > 0");
 
     ft_set_cur_cell(*table, row, col);
 
@@ -108,16 +110,29 @@ static int lft_row_count(lua_State *L) {
     ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
     size_t row_count = ft_row_count(*table);
 
-    lua_pushnumber(L, row_count);
+    lua_pushinteger(L, row_count);
+    return 1;
+}
+
+static int lft_col_count(lua_State *L) {
+    ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
+    size_t col_count = ft_col_count(*table);
+
+    lua_pushinteger(L, col_count);
     return 1;
 }
 
 static int lft_erase_range(lua_State *L) {
     ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
-    size_t top_left_row = luaL_checknumber(L, 2);
-    size_t top_left_col = luaL_checknumber(L, 3);
-    size_t bottom_right_row = luaL_checknumber(L, 4);
-    size_t bottom_right_col = luaL_checknumber(L, 5);
+    size_t top_left_row = luaL_checknumber(L, 2) - 1;
+    size_t top_left_col = luaL_checknumber(L, 3) - 1;
+    size_t bottom_right_row = luaL_checknumber(L, 4) - 1;
+    size_t bottom_right_col = luaL_checknumber(L, 5) - 1;
+
+    luaL_argcheck(L, lua_tonumber(L, 2) > 0, 2, "must be > 0");
+    luaL_argcheck(L, lua_tonumber(L, 3) > 0, 3, "must be > 0");
+    luaL_argcheck(L, lua_tonumber(L, 4) > 0, 4, "must be > 0");
+    luaL_argcheck(L, lua_tonumber(L, 5) > 0, 5, "must be > 0");
 
     ERR_CHECK(ft_erase_range(*table, top_left_row, top_left_col,
                              bottom_right_row, bottom_right_col));
@@ -199,8 +214,6 @@ static int lft_to_string(lua_State *L) {
     return 1;
 }
 
-static int lft_create_border_style(lua_State *L) { return 0; }
-
 static int lft_set_default_border_style(lua_State *L) {
     const struct ft_border_style **style = luaL_checkudata(L, 1, FBORDERSTYLE);
 
@@ -219,15 +232,23 @@ static int lft_set_default_cell_prop(lua_State *L) {
     uint32_t property = luaL_checknumber(L, 1);
     int value = luaL_checknumber(L, 2);
 
+    luaL_argcheck(L, lua_tonumber(L, 1) > 0, 1, "must be > 0");
+
     ERR_CHECK(ft_set_default_cell_prop(property, value));
     return 0;
 }
 static int lft_set_cell_prop(lua_State *L) {
     ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
     size_t row = luaL_checknumber(L, 2);
+    row -= (row > FT_MAX_ROW_INDEX) ? 0 : 1;
     size_t col = luaL_checknumber(L, 3);
+    col -= (col > FT_MAX_COL_INDEX) ? 0 : 1;
     uint32_t property = luaL_checknumber(L, 4);
     int value = luaL_checknumber(L, 5);
+
+    luaL_argcheck(L, lua_tonumber(L, 2) > 0, 2, "must be > 0");
+    luaL_argcheck(L, lua_tonumber(L, 3) > 0, 3, "must be > 0");
+    luaL_argcheck(L, lua_tonumber(L, 4) > 0, 4, "must be > 0");
 
     ERR_CHECK(ft_set_cell_prop(*table, row, col, property, value));
     return 0;
@@ -237,6 +258,8 @@ static int lft_set_default_tbl_prop(lua_State *L) {
     uint32_t property = luaL_checknumber(L, 1);
     int value = luaL_checknumber(L, 2);
 
+    luaL_argcheck(L, lua_tonumber(L, 1) > 0, 1, "must be > 0");
+
     ERR_CHECK(ft_set_default_tbl_prop(property, value));
     return 0;
 }
@@ -245,14 +268,20 @@ static int lft_set_tbl_prop(lua_State *L) {
     uint32_t property = luaL_checknumber(L, 2);
     int value = luaL_checknumber(L, 3);
 
+    luaL_argcheck(L, lua_tonumber(L, 2) > 0, 2, "must be > 0");
+
     ERR_CHECK(ft_set_tbl_prop(*table, property, value));
     return 0;
 }
 static int lft_set_cell_span(lua_State *L) {
     ft_table_t **table = luaL_checkudata(L, 1, FTABLEMETA);
-    size_t row = luaL_checknumber(L, 2);
-    size_t col = luaL_checknumber(L, 3);
+    size_t row = luaL_checknumber(L, 2) - 1;
+    size_t col = luaL_checknumber(L, 3) - 1;
     size_t hor_span = luaL_checknumber(L, 4);
+
+    luaL_argcheck(L, lua_tonumber(L, 2) > 0, 2, "must be > 0");
+    luaL_argcheck(L, lua_tonumber(L, 3) > 0, 3, "must be > 0");
+    luaL_argcheck(L, lua_tonumber(L, 4) > 0, 4, "must be > 0");
 
     ERR_CHECK(ft_set_cell_span(*table, row, col, hor_span));
     return 0;
@@ -267,6 +296,7 @@ static const struct luaL_Reg fort_functions[] = {
     {"set_cur_cell", lft_set_cur_cell},
     {"is_empty", lft_is_empty},
     {"row_count", lft_row_count},
+    {"col_count", lft_col_count},
     {"erase_range", lft_erase_range},
     {"row_write", lft_row_write},
     {"row_write_ln", lft_row_write_ln},
@@ -284,6 +314,7 @@ static const struct luaL_Reg fort_functions[] = {
 
 static const struct luaL_Reg ftable_meta[] = {
     {"__gc", ftable_destroy},
+    {"__tostring", lft_to_string},
     {NULL, NULL},
 };
 
@@ -311,6 +342,7 @@ int luaopen_cfort(lua_State *L) {
 #else
     luaL_openlib(L, 0, ftable_meta, 0);
 #endif
+
     // add fort functions to ftable object
     lua_pushliteral(L, "__index");
     lua_pushvalue(L, -3);  // dup methods table
@@ -341,127 +373,127 @@ int luaopen_cfort(lua_State *L) {
     REGISTER_FORT_STYLE("FRAME_STYLE", FT_FRAME_STYLE);
 
     // special macros
-    lua_pushnumber(L, FT_ANY_COLUMN);
+    lua_pushinteger(L, FT_ANY_COLUMN);
     lua_setfield(L, -2, "ANY_COLUMN");
-    lua_pushnumber(L, FT_CUR_COLUMN);
+    lua_pushinteger(L, FT_CUR_COLUMN);
     lua_setfield(L, -2, "CUR_COLUMN");
-    lua_pushnumber(L, FT_ANY_ROW);
+    lua_pushinteger(L, FT_ANY_ROW);
     lua_setfield(L, -2, "ANY_ROW");
-    lua_pushnumber(L, FT_CUR_ROW);
+    lua_pushinteger(L, FT_CUR_ROW);
     lua_setfield(L, -2, "CUR_ROW");
 
     // cell properties
-    lua_pushnumber(L, FT_CPROP_MIN_WIDTH);
+    lua_pushinteger(L, FT_CPROP_MIN_WIDTH);
     lua_setfield(L, -2, "CPROP_MIN_WIDTH");
-    lua_pushnumber(L, FT_CPROP_TEXT_ALIGN);
+    lua_pushinteger(L, FT_CPROP_TEXT_ALIGN);
     lua_setfield(L, -2, "CPROP_TEXT_ALIGN");
-    lua_pushnumber(L, FT_CPROP_TOP_PADDING);
+    lua_pushinteger(L, FT_CPROP_TOP_PADDING);
     lua_setfield(L, -2, "CPROP_TOP_PADDING");
-    lua_pushnumber(L, FT_CPROP_BOTTOM_PADDING);
+    lua_pushinteger(L, FT_CPROP_BOTTOM_PADDING);
     lua_setfield(L, -2, "CPROP_BOTTOM_PADDING");
-    lua_pushnumber(L, FT_CPROP_LEFT_PADDING);
+    lua_pushinteger(L, FT_CPROP_LEFT_PADDING);
     lua_setfield(L, -2, "CPROP_LEFT_PADDING");
-    lua_pushnumber(L, FT_CPROP_RIGHT_PADDING);
+    lua_pushinteger(L, FT_CPROP_RIGHT_PADDING);
     lua_setfield(L, -2, "CPROP_RIGHT_PADDING");
-    lua_pushnumber(L, FT_CPROP_EMPTY_STR_HEIGHT);
+    lua_pushinteger(L, FT_CPROP_EMPTY_STR_HEIGHT);
     lua_setfield(L, -2, "CPROP_EMPTY_STR_HEIGHT");
-    lua_pushnumber(L, FT_CPROP_ROW_TYPE);
+    lua_pushinteger(L, FT_CPROP_ROW_TYPE);
     lua_setfield(L, -2, "CPROP_ROW_TYPE");
-    lua_pushnumber(L, FT_CPROP_CONT_FG_COLOR);
+    lua_pushinteger(L, FT_CPROP_CONT_FG_COLOR);
     lua_setfield(L, -2, "CPROP_CONT_FG_COLOR");
-    lua_pushnumber(L, FT_CPROP_CELL_BG_COLOR);
+    lua_pushinteger(L, FT_CPROP_CELL_BG_COLOR);
     lua_setfield(L, -2, "CPROP_CELL_BG_COLOR");
-    lua_pushnumber(L, FT_CPROP_CONT_BG_COLOR);
+    lua_pushinteger(L, FT_CPROP_CONT_BG_COLOR);
     lua_setfield(L, -2, "CPROP_CONT_BG_COLOR");
-    lua_pushnumber(L, FT_CPROP_CELL_TEXT_STYLE);
+    lua_pushinteger(L, FT_CPROP_CELL_TEXT_STYLE);
     lua_setfield(L, -2, "CPROP_CELL_TEXT_STYLE");
-    lua_pushnumber(L, FT_CPROP_CONT_TEXT_STYLE);
+    lua_pushinteger(L, FT_CPROP_CONT_TEXT_STYLE);
     lua_setfield(L, -2, "CPROP_CONT_TEXT_STYLE");
 
     // colors
-    lua_pushnumber(L, FT_COLOR_DEFAULT);
+    lua_pushinteger(L, FT_COLOR_DEFAULT);
     lua_setfield(L, -2, "COLOR_DEFAULT");
-    lua_pushnumber(L, FT_COLOR_BLACK);
+    lua_pushinteger(L, FT_COLOR_BLACK);
     lua_setfield(L, -2, "COLOR_BLACK");
-    lua_pushnumber(L, FT_COLOR_RED);
+    lua_pushinteger(L, FT_COLOR_RED);
     lua_setfield(L, -2, "COLOR_RED");
-    lua_pushnumber(L, FT_COLOR_GREEN);
+    lua_pushinteger(L, FT_COLOR_GREEN);
     lua_setfield(L, -2, "COLOR_GREEN");
-    lua_pushnumber(L, FT_COLOR_YELLOW);
+    lua_pushinteger(L, FT_COLOR_YELLOW);
     lua_setfield(L, -2, "COLOR_YELLOW");
-    lua_pushnumber(L, FT_COLOR_BLUE);
+    lua_pushinteger(L, FT_COLOR_BLUE);
     lua_setfield(L, -2, "COLOR_BLUE");
-    lua_pushnumber(L, FT_COLOR_MAGENTA);
+    lua_pushinteger(L, FT_COLOR_MAGENTA);
     lua_setfield(L, -2, "COLOR_MAGENTA");
-    lua_pushnumber(L, FT_COLOR_CYAN);
+    lua_pushinteger(L, FT_COLOR_CYAN);
     lua_setfield(L, -2, "COLOR_CYAN");
-    lua_pushnumber(L, FT_COLOR_LIGHT_GRAY);
+    lua_pushinteger(L, FT_COLOR_LIGHT_GRAY);
     lua_setfield(L, -2, "COLOR_LIGHT_GRAY");
-    lua_pushnumber(L, FT_COLOR_DARK_GRAY);
+    lua_pushinteger(L, FT_COLOR_DARK_GRAY);
     lua_setfield(L, -2, "COLOR_DARK_GRAY");
-    lua_pushnumber(L, FT_COLOR_LIGHT_RED);
+    lua_pushinteger(L, FT_COLOR_LIGHT_RED);
     lua_setfield(L, -2, "COLOR_LIGHT_RED");
-    lua_pushnumber(L, FT_COLOR_LIGHT_GREEN);
+    lua_pushinteger(L, FT_COLOR_LIGHT_GREEN);
     lua_setfield(L, -2, "COLOR_LIGHT_GREEN");
-    lua_pushnumber(L, FT_COLOR_LIGHT_YELLOW);
+    lua_pushinteger(L, FT_COLOR_LIGHT_YELLOW);
     lua_setfield(L, -2, "COLOR_LIGHT_YELLOW");
-    lua_pushnumber(L, FT_COLOR_LIGHT_BLUE);
+    lua_pushinteger(L, FT_COLOR_LIGHT_BLUE);
     lua_setfield(L, -2, "COLOR_LIGHT_BLUE");
-    lua_pushnumber(L, FT_COLOR_LIGHT_MAGENTA);
+    lua_pushinteger(L, FT_COLOR_LIGHT_MAGENTA);
     lua_setfield(L, -2, "COLOR_LIGHT_MAGENTA");
-    lua_pushnumber(L, FT_COLOR_LIGHT_CYAN);
+    lua_pushinteger(L, FT_COLOR_LIGHT_CYAN);
     lua_setfield(L, -2, "COLOR_LIGHT_CYAN");
-    lua_pushnumber(L, FT_COLOR_LIGHT_WHYTE);
+    lua_pushinteger(L, FT_COLOR_LIGHT_WHYTE);
     lua_setfield(L, -2, "COLOR_LIGHT_WHYTE");
 
     // text style
-    lua_pushnumber(L, FT_TSTYLE_DEFAULT);
+    lua_pushinteger(L, FT_TSTYLE_DEFAULT);
     lua_setfield(L, -2, "TSTYLE_DEFAULT");
-    lua_pushnumber(L, FT_TSTYLE_BOLD);
+    lua_pushinteger(L, FT_TSTYLE_BOLD);
     lua_setfield(L, -2, "TSTYLE_BOLD");
-    lua_pushnumber(L, FT_TSTYLE_DIM);
+    lua_pushinteger(L, FT_TSTYLE_DIM);
     lua_setfield(L, -2, "TSTYLE_DIM");
-    lua_pushnumber(L, FT_TSTYLE_ITALIC);
+    lua_pushinteger(L, FT_TSTYLE_ITALIC);
     lua_setfield(L, -2, "TSTYLE_ITALIC");
-    lua_pushnumber(L, FT_TSTYLE_UNDERLINED);
+    lua_pushinteger(L, FT_TSTYLE_UNDERLINED);
     lua_setfield(L, -2, "TSTYLE_UNDERLINED");
-    lua_pushnumber(L, FT_TSTYLE_BLINK);
+    lua_pushinteger(L, FT_TSTYLE_BLINK);
     lua_setfield(L, -2, "TSTYLE_BLINK");
-    lua_pushnumber(L, FT_TSTYLE_INVERTED);
+    lua_pushinteger(L, FT_TSTYLE_INVERTED);
     lua_setfield(L, -2, "TSTYLE_INVERTED");
-    lua_pushnumber(L, FT_TSTYLE_HIDDEN);
+    lua_pushinteger(L, FT_TSTYLE_HIDDEN);
     lua_setfield(L, -2, "TSTYLE_HIDDEN");
 
     // text alignment
-    lua_pushnumber(L, FT_ALIGNED_LEFT);
+    lua_pushinteger(L, FT_ALIGNED_LEFT);
     lua_setfield(L, -2, "ALIGNED_LEFT");
-    lua_pushnumber(L, FT_ALIGNED_CENTER);
+    lua_pushinteger(L, FT_ALIGNED_CENTER);
     lua_setfield(L, -2, "ALIGNED_CENTER");
-    lua_pushnumber(L, FT_ALIGNED_RIGHT);
+    lua_pushinteger(L, FT_ALIGNED_RIGHT);
     lua_setfield(L, -2, "ALIGNED_RIGHT");
 
     // row type
-    lua_pushnumber(L, FT_ROW_COMMON);
+    lua_pushinteger(L, FT_ROW_COMMON);
     lua_setfield(L, -2, "ROW_COMMON");
-    lua_pushnumber(L, FT_ROW_HEADER);
+    lua_pushinteger(L, FT_ROW_HEADER);
     lua_setfield(L, -2, "ROW_HEADER");
 
     // text prop
-    lua_pushnumber(L, FT_TPROP_LEFT_MARGIN);
+    lua_pushinteger(L, FT_TPROP_LEFT_MARGIN);
     lua_setfield(L, -2, "TPROP_LEFT_MARGIN");
-    lua_pushnumber(L, FT_TPROP_TOP_MARGIN);
+    lua_pushinteger(L, FT_TPROP_TOP_MARGIN);
     lua_setfield(L, -2, "TPROP_TOP_MARGIN");
-    lua_pushnumber(L, FT_TPROP_RIGHT_MARGIN);
+    lua_pushinteger(L, FT_TPROP_RIGHT_MARGIN);
     lua_setfield(L, -2, "TPROP_RIGHT_MARGIN");
-    lua_pushnumber(L, FT_TPROP_BOTTOM_MARGIN);
+    lua_pushinteger(L, FT_TPROP_BOTTOM_MARGIN);
     lua_setfield(L, -2, "TPROP_BOTTOM_MARGIN");
-    lua_pushnumber(L, FT_TPROP_ADDING_STRATEGY);
+    lua_pushinteger(L, FT_TPROP_ADDING_STRATEGY);
     lua_setfield(L, -2, "TPROP_ADDING_STRATEGY");
 
     // adding strategy
-    lua_pushnumber(L, FT_STRATEGY_REPLACE);
+    lua_pushinteger(L, FT_STRATEGY_REPLACE);
     lua_setfield(L, -2, "STRATEGY_REPLACE");
-    lua_pushnumber(L, FT_STRATEGY_INSERT);
+    lua_pushinteger(L, FT_STRATEGY_INSERT);
     lua_setfield(L, -2, "STRATEGY_INSERT");
 
     return 1;